+++ docs/changelog.rst
@@ -8,9 +8,10 @@ v0.3.0b1
 
 - 移除对 Python2 的支持
 - 新增计数模块 `datasets.section_counter`
-- 部分函数增加 Key-Only Arguments (`PEP3102`_) 限定
+- 部分函数增加 Key-Only Arguments (`PEP 3102`_) 限定
+- 发布数据构建模块文档
 
-.. _PEP3102: https://www.python.org/dev/peps/pep-3102/
+.. _PEP 3102: https://www.python.org/dev/peps/pep-3102/
 
 v0.2.3 (20180114)
 -----------------